requirements.txt
asgiref==3.3.4
certifi==2021.5.30
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions==3.10.0.0
wincertstore==0.2
